requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.6
psycopg2-binary>=2.9
dj-database-url>=2.1
boto3>=1.34
django-storages>=1.14
gunicorn>=21.2
Pillow>=10.0
